change label tensor type from long to float32 in ECGDataset

+++ src/torch_utilities.py
@@ -202,7 +202,7 @@ class ECGDataset(Dataset):
             sample = self.transform(sample)
         # Convert to torch.Tensor
         sample = torch.tensor(sample, dtype=torch.float32)
-        label = torch.tensor(label, dtype=torch.long)
+        label = torch.tensor(label, dtype=torch.float32)
         return sample, label
